ai_engine/requirements.txt
confluent-kafka
python-dotenv
google-cloud-aiplatform